Implement MainViewController to show the UI of the Application
In order to show the UI for the application we needed to add a new swift file and xib called MainViewController in which the application's AppDelegate will instantiate or call in order to assign it to the appDelegate's NSWindow object.  The NSWindow object is the main window of any Cocoa or Cocoa Touch application that presents the UI to the user.  Having MainViewController is a requirement.