Add a text field to the MainViewController xib for the UI
Added a NSTextField object as a child to the MainViewController window.  The NSTextField object has a boolean variable where you can set it to true of false for 'editable'.  This variable on this NSTextField object will be set to false as we only want user to be able to copy and past but not edit the text.  We actually set this variable from the storyboard's attributes and not in code.